Burner On Visualizations (#9)
Added visualizations for when burner is on without any pot. The gif shows some possible transitions -- if a pot is added, it transitions to a proactive timer. If the pot is again removed, it goes back to the burner on visualization. 

Once the burner is turned off it fades away. 

![2019-04-02 16 26 11](https://user-images.githubusercontent.com/19956321/55433945-78020680-5564-11e9-97e0-b92825fa0f1e.gif)

I also added two features into ImageTrackerLerper. One that decreases the lerp speed as distance to the target increases and another that prevents objects from being clipped by the near clipping plane.

In addition, I fixed a number of race conditions prevalent with voice input.